requirements.txt
urllib3=1.26.15
mistralai>=1.10.0
python-dotenv>=1.2.1
openpyxl>=3.1.5
pandas>=2.0.0
xlrd>=2.0.1
